+++ src/meltano/cli/discovery.py
@@ -1,14 +1,9 @@
-import json
-import os
-from urllib.parse import urlparse
+"""Discoverable Plugins CLI."""
 
 import click
+
 from meltano.core.plugin import PluginType
-from meltano.core.plugin_discovery_service import (
-    DiscoveryInvalidError,
-    PluginDiscoveryService,
-)
-from meltano.core.project import Project
+from meltano.core.plugin_discovery_service import PluginDiscoveryService
 from meltano.core.tracking import GoogleAnalyticsTracker
 
 from . import cli
@@ -24,7 +19,7 @@ def discover(project, plugin_type):
     """
     List the available discoverable plugins and their variants.
 
-    \b\nRead more at https://meltano.com/docs/command-line-interface.html#discover
+    \b\nRead more at https://docs.meltano.com/reference/command-line-interface#discover
     """
     discover_service = PluginDiscoveryService(project)
     if plugin_type == "all":
@@ -32,13 +27,13 @@ def discover(project, plugin_type):
     else:
         plugin_types = [PluginType.from_cli_argument(plugin_type)]
 
-    for i, plugin_type in enumerate(plugin_types):
-        if i > 0:
+    for idx, discovered_plugin_type in enumerate(plugin_types):
+        if idx > 0:
             click.echo()
 
-        click.secho(f"{str(plugin_type).capitalize()}", fg="green")
+        click.secho(f"{str(discovered_plugin_type).capitalize()}", fg="green")
 
-        for plugin_def in discover_service.get_plugins_of_type(plugin_type):
+        for plugin_def in discover_service.get_plugins_of_type(discovered_plugin_type):
             click.echo(plugin_def.name, nl=False)
 
             if len(plugin_def.variants) > 1: