+++ src/meltano/core/meltano_file.py
@@ -1,5 +1,6 @@
+"""Module for working with meltano.yml files."""
 import copy
-from typing import Dict, Iterable, List
+from typing import Dict, Iterable, List, Optional
 
 from meltano.core.behavior.canonical import Canonical
 from meltano.core.environment import Environment
@@ -11,25 +12,46 @@ VERSION = 1
 
 
 class MeltanoFile(Canonical):
+    """Data and loading methods for meltano.yml files."""
+
     def __init__(
         self,
         version: int = VERSION,
+        default_environment: Optional[str] = None,
         plugins: Dict[str, dict] = None,
         schedules: List[dict] = None,
         environments: List[dict] = None,
         **extras,
     ):
+        """Construct a new MeltanoFile object from meltano.yml file.
+
+        Args:
+            version: The meltano.yml version, currently always 1.
+            default_environment: The default environment to use for commands in this project.
+            plugins: Plugin configuration for this project.
+            schedules: Schedule configuration for this project.
+            environments: Environment configuration for this project.
+            extras: Additional configuration for this project.
+        """
         super().__init__(
             # Attributes will be listed in meltano.yml in this order:
             version=version,
+            default_environment=default_environment,
             extras=extras,
             plugins=self.load_plugins(plugins or {}),
             schedules=self.load_schedules(schedules or []),
             environments=self.load_environments(environments or []),
         )
 
-    def load_plugins(self, plugins) -> Canonical:
-        """Parse the meltano.yml file and return it as `ProjectPlugin` instances."""
+    def load_plugins(self, plugins: Dict[str, dict]) -> Canonical:
+        """Parse the meltano.yml file and return it as `ProjectPlugin` instances.
+
+        Args:
+            plugins: Dict of plugin configurations.
+
+        Returns:
+            New ProjectPlugin instances.
+        """
         plugin_type_plugins = Canonical()
 
         for ptype in PluginType:
@@ -53,7 +75,15 @@ class MeltanoFile(Canonical):
 
         return plugin_type_plugins
 
-    def load_schedules(self, schedules) -> List[Schedule]:
+    def load_schedules(self, schedules: List[dict]) -> List[Schedule]:
+        """Parse the meltano.yml file and return it as Schedule instances.
+
+        Args:
+            schedules: List of schedule configurations.
+
+        Returns:
+            List of new Schedule instances.
+        """
         return list(map(Schedule.parse, schedules))
 
     @staticmethod
@@ -74,6 +104,7 @@ class MeltanoFile(Canonical):
 
         Args:
             mapper_config: The dict representation of a mapper config found in in meltano.yml.
+
         Returns:
             A list of `ProjectPlugin` instances.
         """