+++ tests/meltano/cli/test_cli.py
@@ -1,9 +1,8 @@
-import os
 import shutil
-from copy import copy
+
+import pytest
 
 import meltano
-import pytest
 from meltano.cli import cli
 from meltano.core.project import PROJECT_READONLY_ENV, Project
 from meltano.core.project_settings_service import ProjectSettingsService
@@ -12,13 +11,19 @@ from meltano.core.project_settings_service import ProjectSettingsService
 class TestCli:
     @pytest.fixture()
     def project(self, test_dir, project_init_service):
-        """This fixture returns the non-activated project."""
-        project = project_init_service.init(activate=False, add_discovery=True)
+        """Return the non-activated project."""
+        project = project_init_service.init(  # noqa: DAR301
+            activate=False, add_discovery=True
+        )
 
         yield project
 
         Project.deactivate()
         shutil.rmtree(project.root)
+
+    @pytest.fixture()
+    def deactivate_project(self):
+        Project.deactivate()
 
     def test_activate_project(self, project, cli_runner, pushd):
         assert Project._default is None
@@ -28,7 +33,7 @@ class TestCli:
 
         assert Project._default is not None
         assert Project._default.root == project.root
-        assert Project._default.readonly == False
+        assert Project._default.readonly is False
 
     def test_activate_project_readonly_env(
         self, project, cli_runner, pushd, monkeypatch
@@ -40,7 +45,7 @@ class TestCli:
         pushd(project.root)
         cli_runner.invoke(cli, ["discover"])
 
-        assert Project._default.readonly == True
+        assert Project._default.readonly
 
     def test_activate_project_readonly_dotenv(
         self, project, cli_runner, pushd, monkeypatch
@@ -52,9 +57,83 @@ class TestCli:
         pushd(project.root)
         cli_runner.invoke(cli, ["discover"])
 
-        assert Project._default.readonly == True
+        assert Project._default.readonly
 
     def test_version(self, cli_runner):
         cli_version = cli_runner.invoke(cli, ["--version"])
 
         assert cli_version.output == f"meltano, version {meltano.__version__}\n"
+
+    def test_default_environment_is_activated(
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["discover"],
+        )
+        assert Project._default.active_environment.name == "test-meltano-environment"
+
+    def test_environment_flag_overrides_default(
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["--environment", "test-subconfig-2-yml", "discover"],
+        )
+
+        assert Project._default.active_environment.name == "test-subconfig-2-yml"
+
+    def test_environment_variable_overrides_default(
+        self, deactivate_project, project_files_cli, cli_runner, pushd, monkeypatch
+    ):
+
+        monkeypatch.setenv("MELTANO_ENVIRONMENT", "test-subconfig-2-yml")
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["discover"],
+        )
+        assert Project._default.active_environment.name == "test-subconfig-2-yml"
+
+    def test_lower_null_environment_overrides_default(
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["--environment", "null", "discover"],
+        )
+        assert Project._default.active_environment is None
+
+    def test_upper_null_environment_overrides_default(
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["--environment", "NULL", "discover"],
+        )
+        assert Project._default.active_environment is None
+
+    def test_no_environment_overrides_default(
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["--no-environment", "discover"],
+        )
+        assert Project._default.active_environment is None
+
+    def test_no_environment_and_null_environment_overrides_default(  # noqa: WPS118
+        self, deactivate_project, project_files_cli, cli_runner, pushd
+    ):
+        pushd(project_files_cli.root)
+        cli_runner.invoke(
+            cli,
+            ["--no-environment", "--environment", "null", "discover"],
+        )
+        assert Project._default.active_environment is None